requirements.txt
pymupdf==1.24.2 
pytesseract==0.3.10 
pillow==10.4.0 
nltk==3.8.1 
sentence-transformers==2.7.0 
psycopg2-binary==2.9.9 
pgvector==0.2.5 
fastapi==0.115.0 
pydantic==2.9.2 
redis==5.0.8 
transformers==4.44.2 
loguru==0.7.2 
uvicorn==0.30.6 
pytest==8.3.2 
flake8==7.1.1
sentencepiece